+++ Pandora/Paradigm/Structure/Some/List.hs
@@ -51,7 +51,7 @@ import Pandora.Paradigm.Structure.Ability.Morphable (Morphable (Morphing, morphi
 import Pandora.Paradigm.Structure.Ability.Slidable (Slidable (Sliding, slide))
 import Pandora.Paradigm.Structure.Ability.Substructure (Substructure (Substance, substructure, sub), Segment (Root, Rest))
 import Pandora.Paradigm.Structure.Interface.Stack (Stack (Topping, push, pop, top))
-import Pandora.Paradigm.Structure.Interface.Zipper (Zippable (Breadcrumbs, fasten), Zipper)
+import Pandora.Paradigm.Structure.Interface.Zipper (Zippable (Breadcrumbs, fasten, unfasten), Zipper)
 import Pandora.Paradigm.Structure.Modification.Combinative (Combinative)
 import Pandora.Paradigm.Structure.Modification.Comprehension (Comprehension (Comprehension))
 import Pandora.Paradigm.Structure.Modification.Prefixed (Prefixed)
@@ -173,6 +173,10 @@ type instance Combinative List = Comprehension Maybe
 instance Zippable List where
 	type Breadcrumbs List = Reverse List <:*:> List
 	fasten (TT (Just (Construct x xs))) = Just <----- Exactly x <:*:> Reverse <-- TT empty <:*:> TT xs
+	fasten (TT Nothing) = Nothing
+	unfasten :: forall e . Zipper List e -> Nonempty List e
+	unfasten (T_U (Exactly focus :*: T_U (Reverse left :*: right))) =
+		attached <-- (push @(Nonempty List) <-/- left) <~ Construct focus (run right) where
 
 -- TODO: No overlapping, let's use wrappers instead
 instance {-# OVERLAPS #-} Traversable (->) (->) (Tape List) where